+++ src/modules/transportation/entities/spt-requisition.vw.ts
@@ -17,6 +17,14 @@ export class RequisitionView extends BaseEntity {
   destinationName!: string;
 
   @Field()
+  @Column({ name: 'DS' })
+  ds!: string;
+
+  @Field({ nullable: true })
+  @Column({ name: 'DIVISI' })
+  divisi?: string;
+
+  @Field()
   @Column({ name: 'CUSTOMER_ID' })
   customerId!: string;
 
@@ -33,16 +41,20 @@ export class RequisitionView extends BaseEntity {
   rollQty!: number;
 
   @Field({ nullable: true })
+  @Column({ name: 'SPACE' })
+  space?: number;
+
+  @Field({ nullable: true })
   @Column({ name: 'METER' })
   meter?: number;
 
-  @Field()
+  @Field({ nullable: true })
   @Column({ name: 'WEIGHT' })
-  weight!: number;
+  weight?: number;
 
-  @Field()
+  @Field({ nullable: true })
   @Column({ name: 'VOLUME' })
-  volume!: number;
+  volume?: number;
 
   @Field()
   @Column({ name: 'CONTRACT' })